Move term env and cmd tracker to node layer
