+++ src/vs/workbench/parts/terminal/node/terminalEnvironment.ts
