+++ src/vs/workbench/parts/terminal/test/node/terminalCommandTracker.test.ts
