+++ src/vs/workbench/parts/terminal/test/node/terminalEnvironment.test.ts
@@ -6,7 +6,7 @@
 import * as assert from 'assert';
 import * as os from 'os';
 import * as platform from 'vs/base/common/platform';
-import * as terminalEnvironment from 'vs/workbench/parts/terminal/electron-browser/terminalEnvironment';
+import * as terminalEnvironment from 'vs/workbench/parts/terminal/node/terminalEnvironment';
 import Uri from 'vs/base/common/uri';
 import { IStringDictionary } from 'vs/base/common/collections';
 import { IShellLaunchConfig, ITerminalConfigHelper } from 'vs/workbench/parts/terminal/common/terminal';
